backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
motor==3.6.0
pymongo>=4.7,<4.10
redis==5.0.0
google-genai==1.5.0
anthropic==0.39.0
openai==1.50.0
httpx>=0.28.1
python-multipart==0.0.9
Pillow==10.4.0
pydantic==2.9.0
